refactored to give job and company functions their own file
