+++ app/core/database.php
@@ -2,9 +2,6 @@
 
 class Database
 {
-    public function __construct()
-    {
-    }
     public static function connectDb()
     {
         try {
@@ -19,43 +16,5 @@ class Database
             echo "Please contact the administrator. <br>";
             die("Connection failed: " . $e->getMessage());
         }
-    }
-
-    public function getJobs()
-    {
-        $pdo = self::connectDb();
-        $stmt = $pdo->prepare("
-            SELECT jobs.*, companies.name as company 
-            FROM jobs 
-            LEFT JOIN companies ON jobs.company_id = companies.id
-        ");
-        $stmt->execute();
-        return $stmt->fetchAll();
-    }
-    
-    public function getCompanies()
-    {
-        $pdo = self::connectDb();
-        $stmt = $pdo->prepare("SELECT * FROM companies");
-        $stmt->execute();
-        return $stmt->fetchAll();
-    }
-
-    public function filterJobs($filters)
-    {
-        $pdo = self::connectDb();
-        // Check if $filters is an array
-        if (!is_array($filters)) {
-            $filters = [$filters];
-        }
-        // Create a string of placeholders for each filter value
-        $filterPlaceholders = implode(',', array_fill(0, count($filters), '?'));
-        $stmt = $pdo->prepare("SELECT * FROM jobs WHERE category_id IN ($filterPlaceholders)");
-        
-        foreach ($filters as $index => $filter) {
-            $stmt->bindValue($index + 1, $filter, PDO::PARAM_INT);
-        }
-        $stmt->execute();
-        return $stmt->fetchAll();
     }
 }
